refactor(backend): typed-buffer foundation — shared Dtype enum + delete dead argmax (PR A)
Phase A of the typed-buffer cleanup planned in this thread (Backend
trait wart cleanup, sibling to PRs #157-#159). Pure foundation: extend
the dtype tag enum to cover the integer dtypes currently tunneled
through `Self::Buffer` via `alloc_u32` / `write_u32` / `from_slice_i32`,
move it to a shared module so CPU/CUDA/Metal can all consume it, and
delete the never-called `argmax_rows_to_u32` trait method + matching
.cu kernel.

Concretely:

- New `crates/ferrum-kernels/src/backend/dtype.rs` — `Dtype` enum with
  F32 / F16 / U32 / I32 / I8. `bytes_per_elem()` + `name()` helpers.
- Metal's local `Dtype` (F32/F16-only) replaced by a re-export of the
  shared enum. `MetalBuf` keeps using it via the same path. Existing
  match arms get a panicking wildcard for the new integer variants —
  no caller produces an integer MetalBuf yet (that's Phase B's job),
  so reaching the panic = caller bug.
- `argmax_rows_to_u32` removed from `Backend` trait + CUDA override +
  `kernels/argmax_rows.cu` deleted + `build.rs` PTX entry removed.
  Zero callsites in the workspace — confirmed by `grep -rn`. Was a
  layering violation anyway (sampling decision on Backend, not in
  `ferrum-sampler`).

Phase B (next PR): migrate `Self::Buffer` to typed wrappers
(`CpuBuf` / `CudaBuf` matching Metal's pattern), replace the
`alloc_u32` / `write_u32` / `from_slice_i32` family with a unified
`Self::alloc(Dtype, n)` / `Self::write(buf, &[T])`. ~15-20 file
touches, will need Metal+CUDA bench validation before merge.

## Test plan

- [x] `cargo check --workspace --features metal` clean (Mac)
- [x] `cargo check --workspace --no-default-features` clean
- [x] `cargo test --workspace --no-default-features` — full pass
- [x] Metal smoke Qwen3-0.6B c=1 = 59.1 tok/s (vs PR #159 baseline 58.2)
- [ ] CUDA cargo check — pod offline at submission. PR A is type-system
  clean for CUDA (only removed `argmax_rows_to_u32` override + added
  panicking match arms; no impl signature changes).
